+++ src/main/java/com/eiv/interfaces/IPrestamo.java
@@ -7,7 +7,7 @@ public interface IPrestamo {
 
     public Long getNumeroDocumento();
     
-    public Integer getTipoDocumentoId();
+    public Long getTipoDocumentoId();
     
     public LocalDate getFechaAlta();
